src/backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Environment Configuration
python-dotenv==1.0.0

# System Information
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Additional utilities
requests==2.31.0
